wait-for-root: Update for libudev queue API change
libudev no longer exposes the list of devices currently queued.  This
change was made in version 213 by commit 9ea28c55a248 ("udev: remove
seqnum API and all assumptions about seqnums") because the API is
broken in the presence of device namespaces.

Instead of checking whether a particular device is in the queue,
check whether the queue is non-empty.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>