update-initramfs: Run 'sync' after writing the initramfs
Bug #783620 involves several machines with root on xfs being upgraded
from wheezy to jessie and then immediately rebooted.  In some cases
the initramfs was apparently written to the journal but not committed
during the reboot.  As GRUB doesn't replay the journal, it loaded an
empty initramfs and the kernel failed to boot.

There is clearly a bug elsewhere (init system or kernel) that
prevented the filesystem being cleanly shut down.  However, there
seems to be a general risk here, so add a 'sync' to mitigate it.

Closes: #783620
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
(cherry picked from commit fa1985b91f40fc2c3890e84320c61eb2dc2ce9c9)