scripts/functions: Connect panic shell to normal tty device
Currently the panic shell's controlling tty is /dev/console which is
not fully functional - the shell can't provide job control and more
can't work out the screen size for paging.

Fix this by reading /proc/consoles to find out the underlying tty
device and then connecting the shell to it directly with the aid of
setsid.

Closes: #512679
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>