init: Take over handling of rootdelay parameter from udev boot script
The rootdelay parameter used to be important as a workaround for our
crappy device scanning.  It's less widely needed now but still needed
for some md-RAID configurations.

The delay used to be implemented in udev's init-top script, but was
removed from there in Debian version 220-4 and the systemd maintainers
do not want to restore it.  So implement it in init, where it probably
always should have been.

I would like to un-export ROOTDELAY now, but several other packages
use it to control their own delay or polling.  For now, document it as
deprecated.  To avoid doubling the delay when using an older udev,
make mkinitramfs patch it out of the udev init-top script.

Closes: #809740
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>