Move initramfs.conf around in maintainer scripts to avoid it being 'obsolete'
initramfs.conf now belongs to initramfs-tools-core, but on upgrade
from version 0.120 or earlier dpkg remembers that it belonged to
initramfs-tools and treats it as 'obsolete' for that package.  This
may confuse administrators into deleting it.

Change/add maintainer scripts to move it away before upgrade of
initramfs-tools and to move it when either postinst runs or the
upgrade is aborted.

Closes: #813229
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>