hook-functions: Dereference links when looking up block device number
block_dev_mod_add usually gets a device name that has already been
canonicalised using 'readlink -f', but this is not true in the case of
the resume device.  By default stat(1) uses lstat(2) which will give
us a device number of 0:0 for a symlink!  Rather than adding another
use of 'readlink -f', add the -L option so it dereferences symlinks to
block devices.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>