Fix initramfs.conf move to ensure default values are upgraded
When we upgrade from a version earlier than 0.123, initramfs.conf is
transfered from initramfs-tools to initramfs-tools-core and is not
modified.  But if it has not been changed from the default shipped in
the old version, it should be upgraded to the new default.  An upgrade
directly from Debian stable (0.120+deb8u2) to testing (0.125) should
change BUSYBOX=yes to BUSYBOX=auto but does not.

Andreas Beckmann extracted and adapted the code from
dpkg-maintscript-helper for conffile removal and renaming, that has
the necessary checks.  I've simplified that a little bit to reduce
code churn.

Thanks: Andreas Beckmann <anbe@debian.org>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Closes: #822671, #825929