WIP: Use mount program from util-linux instead of busybox/klibc
XXX This will break live-boot and fsprotect since they use
'mount -o move'.  Need to make util-linux's mount support this
option properly, or wrap it.

busybox and klibc are lagging behind in support for mount flags, so
use the standard mount program, assumed to come from util-linux.  This
doesn't increase image size much because most of the code is in
libmount.so which is already pulled in as a dependency of fsck.

- Replace hooks/fsck with hooks/mount, which installs both mount and
  (if needed) fsck programs
- Use 'mount --move' instead of 'mount -o move' in init

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Closes: #845302