unmkinitramfs: Rewrite command syntax
I'm not comfortable with unmkinitramfs passing through everything to
cpio, considering what a sharp tool cpio is.  Instead of that,
recognise and pass through the few specific options that are
important.

Always pass --preserve-modification-time --no-absolute-filenames
--quiet.

Take the initramfs filename as a mandatory argument rather than
expecting a file on stdin, to be more consistent lsinitramfs and
mkinitramfs do.

Take an output directory as a mandatory argument (except when
listing), rather than extracting to the current directory.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>