scripts/local: Apply a single time limit to all device discovery
Instead of counting how many times we wait and poll for each critical
device (root or /usr) to appear, use /proc/uptime to tell how long we
have waited in total (starting from when local_top runs).

This is complicated by the hack mdadm's local-block uses to decide
when 2/3 of the time limit has expired.  Use an even worse hack to
keep it working.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>