scripts/local-premount/resume: Use local_device_setup for resume device
Currently we don't wait for the resume device to appear, and will boot
without resuming if it is too slow to appear (e.g. USB storage or, in
the reported case, an NVMe device!).

Use local_device_setup to wait for the device, the same as we do for
the root and /usr devices.  This also takes care of resolving UUID=
and LABEL= syntax, and adds support for PARTUUID= and PARTLABEL=.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Closes: #854791