Include the axp20x_usb_power module in the base modules list for MODULES=most
With newer kernels (including kernel 4.9 as used in stretch),
initializing the USB host controllers on a number of armhf systems
is not possible without having the "axp20x_usb_power" power supply
driver module available. To be able to have the rootfs on a USB
mass storage device on such systems it is therefore necessary to
include this module in the initramfs.

Closes: #866521
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>