Add mlx4 and mlx5 InfiniBand modules to hidden_dep_add_modules
Mellanox ConnectX architecture is:  mlx4_core is the lower level
PCI driver which register on the PCI id, and protocol specific drivers
are depended on it: mlx4_en - for Ethernet and mlx4_ib for Infiniband.
NIC could have multiple ports which can change their type dynamically.
We use the request_module() call to load the relevant protocol driver
when needed: on loading time or at port type change event.

The mlx4_core and mlx5_core modules are included in the initrd, but not
mlx4_ib and mlx5_ib. Thus the request_module() call will not find these
modules and fail load them. The mlx*_ib module loading will not be
retried.

Therefore add the mlx4_ib and mlx5_ib dependency to the
hidden_dep_add_modules function.

Closes: #871595
Signed-off-by: Benjamin Drung <benjamin.drung@profitbricks.com>