docs: Convert maintainer notes to Markdown format
Alioth's gitweb installation used to serve up the maintainer notes as
text/html, and we linked to that.  Gitlab, however, is written with
some more concern for web security, and won't do that except through
explicitly exporting "pages" to be hosted on a per-team domain.

It doesn't really seem worth the trouble of doing that for the one
HTML file in initramfs-tools.  Instead, convert it to Markdown format
which Gitlab will happily render inline on its repository view.

This will also make the document substantially easier to edit in
future.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>