unmkinitramfs, lsinitramfs: Split multiple early initramfs sections
It is possible for the early initramfs to consist of multiple
sections, each a valid uncompressed cpio archive, for example if
amd64-microcode and intel-microcode are both installed and explicitly
enabled.  Each cpio archive ends with an EOF marker consisting of four
zero bytes and possibly additional zero padding.

The kernel's early microcode loader simply skips the zero bytes, but
cpio will stop.  So whenever we find such a group of zero bytes, we
need to stop parsing cpio headers, run cpio over what we found, and
then repeat the process on the following bytes.

Closes: #886424
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>