mkinitramfs: Create usr-merged filesystem layout
On a usr-merged filesystem /bin, /lib*, /sbin are symlinks to the
corresponding subdirectories of /usr.  When the host filesystem looks
like this, copy_file may be asked to copy both /lib/foo and
/usr/lib/foo.  It's too late at this point to make /lib a symlink
in the initramfs, and copy_file isn't general enough to do that.

Avoid this by setting up the initramfs as usr-merged.  This could
possibly cause file conflicts, but it's unlikely because:

* Policy forbids such conflicts on the host filesystem
* Few executables and libraries that go into the initramfs are
  explicitly installed in /usr on the host filesystem, because
  of historical support for having the init system mount /usr
* mkinitramfs never used to create /usr/bin or /usr/lib

Closes: #874523
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>