update-initramfs: Always take over other initramfs images
In Debian, update-initramfs has always refused to overwrite or delete
an initramfs image that it didn't create, unless the -t option is
used.  On Ubuntu, this check is apparently disabled.  The commit log
doesn't provide much of a clue as to why.  Looking at the the state of
Debian and the linux-2.6 package at the time, I think the reasoning
must have been something like:

(1) Multiple initramfs/initrd generators were available and could be
coinstalled.  The configuration file /etc/kernel-img.conf specified
which one kernel packages should invoke ("ramdisk" variable).

(2) Packages that hooked into initramfs-tools would always run update-
initramfs if it existed, and it would immediately update the
initramfs, because triggers hadn't yet been invented.

(3) To keep the alternate initramfs generators working, update-
initramfs therefore had to either (a) check the configuration file to
find out whether initramfs-tools was meant to be used, or (b) keep
track of which initramfs images it created and therefore "owned".

At this point, (1) is no longer true; the "ramdisk" variable is not
honoured and the initramfs generator to be invoked is chosen by
installing one of several conflicting packages.  I think this means
that (3) no longer follows, and whichever of the automatic initramfs
generator packages is currently installed should be considered to
"own" /boot/initrd.img-<version> for all kernel versions.

This commit only disables the checks.  The state directory and the
state updates are retained for now, in case it becomes necessary to
revert this change.

Closes: #765750, #804295
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>