Remove obsolete update-initramfs state in /var/lib/initramfs-tools
update-initramfs used to keep track of which initramfs images it
created and "owned", to avoid overwriting the output of another
coinstalled initramfs generator.  As discussed in #765750, this is no
longer a concern and the state checks have been removed.

Now that I'm confident this hasn't caused a regression, also remove
the state updates and the state directory.  After this, there is no
turning back: a downgrade will leave update-initramfs owning nothing.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>