scripts/functions: Merge information from BOOTIF/DEVICE and ip parameters
If the BOOTIF parameter or the DEVICE configuration parameter is set,
and the ip parameter is also set but does not specify a device name,
we need to convey the device name to ipconfig as well as the ip
parameter value.

Previously we didn't do this if the value was a colon-separated list.
But now that we can parse the list correctly, it's easy to substitute
the device name into it.

Closes: #721088
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>