scripts/functions: Fix parsing of device name from ip parameter
Parameter expansion is not useful for parsing delimited lists of
arbitrary length.  Instead, set IFS=:, disable wildcard expansion, and
use "set".  Use a separate function, as this makes it easy to change
$IFS temporarily.

Also do the parsing earlier, so we can avoid a redundant invocation of
ipconfig.

This was mentioned in #721088, but it's not the main bug reported.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>