unmkinitramfs, lsinitramfs: Avoid zcat compatibility issue by running gzip
zutils provides an alternate implementation of zcat which has a
different interpretation of the -t option.  This has been fixed in
Debian unstable by installing its zcat under a different name.
However, we can avoid this problem even on systems with the older
version by running gzip instead of zcat.

Closes: #903931
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>