initramfs-tools: Change trigger declaration to "interest-await"
dpkg will only put a triggering package in the triggers-awaited state
if both it and the triggered package used "await".  To allow for
packages that really need to await the initramfs update, explicitly
declare our trigger as "interest-await".  (This is equivalent to the
old "interest", but fixes a lintian warning for the old name.)

Related-to: #822730
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>